# docs/api/swagger.yaml
openapi: 3.0.3
info:
  title: Comic & Novel Digital Platform API
  description: Hệ thống API cung cấp dịch vụ đọc truyện tranh (Comic) và tiểu thuyết (Novel), tích hợp thanh toán VIP và gợi ý nội dung bằng AI.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Phiên bản API v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
          example: Request processed successfully
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: AGE_RESTRICTED
            message:
              type: string
              example: Bạn chưa đủ 18 tuổi để xem nội dung này
  responses:
    Success:
      description: Thành công
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
    Error:
      description: Lỗi từ phía client hoặc nghiệp vụ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication & User
      summary: Đăng ký tài khoản mới
      security: []
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/Error'
  /auth/login:
    post:
      tags:
        - Authentication & User
      summary: Đăng nhập và nhận JWT Token
      security: []
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/Error'
...
  /comics/{slug_or_id}:
    get:
      tags:
        - Content Delivery
      summary: Chi tiết truyện, Tác giả, Age Rating
      security: []
      parameters:
        - in: path
          name: slug_or_id
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/Success'
  /comics/{id}/chapters:
    get:
      tags:
        - Content Delivery
      summary: Danh sách chương (Sắp xếp theo sort_order)
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/Success'
  /chapters/{id}:
    get:
      tags:
        - Content Delivery
      summary: Nội dung chương (Text cho Novel hoặc Metadata cho Comic)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/Success'
...
  /ai/summaries/comic/{id}:
    get:
      tags:
        - Reading & AI
      summary: Lấy bản tóm tắt truyện do AI tạo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/Success'
...
  /admin/chapters/{id}/status:
    patch:
      tags:
        - CMS & Administration
      summary: Duyệt/Từ chối chương (APPROVED/REJECTED)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/Success'
...